+++ source/MaterialXGenGlsl/essl/EsslShaderGenerator.h
