+++ source/MaterialXGenGlsl/essl/EsslSyntax.cpp
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-#include <MaterialXGenGlsl/EsslSyntax.h>
+#include <MaterialXGenGlsl/essl/EsslSyntax.h>
 
 #include <MaterialXGenShader/Shader.h>
 #include <MaterialXGenShader/ShaderGenerator.h>
