+++ source/MaterialXGenGlsl/essl/EsslSyntax.h
